[chore] 닉네임 설정 / 불필요 주석 제거 및 패키지 변경 (#64)
