+++ android/support/src/main/java/com/teamtripdraw/android/support/framework/presentation/event/Event.kt
@@ -1,0 +1,15 @@
+package com.teamtripdraw.android.support.framework.presentation.event
+
+open class Event<out T>(val content: T) {
+    var hasBeenHandled = false
+        private set
+
+    fun getContentIfNotHandled(): T? {
+        return if (hasBeenHandled) {
+            null
+        } else {
+            hasBeenHandled = true
+            content
+        }
+    }
+}
