+++ android/support/src/main/java/com/teamtripdraw/android/support/framework/presentation/event/EventObserver.kt
@@ -1,0 +1,11 @@
+package com.teamtripdraw.android.support.framework.presentation.event
+
+import androidx.lifecycle.Observer
+
+class EventObserver<T>(private val onEventUnhandledContent: (T) -> Unit) : Observer<Event<T>> {
+    override fun onChanged(event: Event<T>?) {
+        event?.getContentIfNotHandled()?.let { value ->
+            onEventUnhandledContent(value)
+        }
+    }
+}
